Add files via upload
this program is about a videogame character similar to subwaysurfers in which there are hurdles with varying heights,the character k needs to overcome these hurdles with intially height=1. To increase height of the character we need to dose of a magic potion.So we need to find the no.of doses required to clear all hurdles.If the height given can overcome all hurdles then return.